Introduction to the Page
Created the very basic layout for the website whilst im still waiting for client to provide me with all the images and details. This helps me set a basic layout to then edit and mess around. also give an idea to the client about the size of everything on the page.